backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
neo4j==5.15.0
groq==0.4.2
pydantic==2.5.2
python-multipart==0.0.6
